+++ packages/superset-ui-preset-chart-xy/src/Line/DefaultLegendItemMarkRenderer.tsx
@@ -1,5 +1,5 @@
 import React, { CSSProperties } from 'react';
-import { LegendItemMarkRendererProps } from '../components/legend2/types';
+import { LegendItemMarkRendererProps } from '../components/legend/types';
 import { LineEncodingConfig } from './Encoder';
 
 const MARK_WIDTH = 12;
