+++ packages/superset-ui-preset-chart-xy/src/components/legend/createRenderLegend.tsx
