--- packages/superset-ui-preset-chart-xy/src/legacy.ts
@@ -1,3 +1,0 @@
-export { default as BoxPlotChartPlugin } from './BoxPlot/legacy';
-export { default as ScatterPlotPlugin } from './ScatterPlot/legacy';
-export { default as LineChartPlugin } from './Line/legacy';
