+++ packages/superset-ui-preset-chart-xy/src/utils/createXYChartLayoutWithTheme.ts
